refactor: move ThemeButton from Header folder to Buttons folder
